pyproject.toml
[project]
name = "multi-agent-research"
version = "0.2.0"
description = "Multi-agent research system using Moonshot Kimi API"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["research_agent*"]
